修改CommentMapperBug以及dto数据格式

+++ reshare-model/src/main/java/com/xktime/model/article/pos/BaseVerifyArticle.java
@@ -10,6 +10,6 @@ public class BaseVerifyArticle implements Serializable {
     private String title;
     private String content;
     private String authorName;
-    private byte status;
+    private int status;
     private long bindId;
 }
