+++ reshare-model/src/main/java/com/xktime/model/article/pos/CrawlerArticle.java
@@ -20,6 +20,6 @@ public class CrawlerArticle extends BaseVerifyArticle {
     private Date publishTime;
     private String url;
     private String origin;
-    private byte status;
+    private int status;
 
 }
